+++ avr-hal-generic/src/port.rs
@@ -207,11 +207,12 @@ impl<PIN: PinOps, MODE: mode::Io> Pin<MODE, PIN> {
         ADC: crate::adc::AdcOps<H>,
         CLOCK: crate::clock::Clock,
     {
-        let new = Pin {
+        let mut new = Pin {
             pin: self.pin,
             _mode: PhantomData,
         };
         adc.enable_pin(&new);
+        unsafe { new.pin.make_input(false) };
         new
     }
 }
@@ -334,7 +335,9 @@ impl<PIN: PinOps> OutputPinV0 for Pin<mode::Output, PIN> {
     }
 }
 
-impl<PIN: PinOps> ErrorType for Pin<mode::Output, PIN> { type Error = core::convert::Infallible; }
+impl<PIN: PinOps> ErrorType for Pin<mode::Output, PIN> {
+    type Error = core::convert::Infallible;
+}
 
 impl<PIN: PinOps> OutputPin for Pin<mode::Output, PIN> {
     fn set_low(&mut self) -> Result<(), Self::Error> {
@@ -439,7 +442,9 @@ impl<PIN: PinOps> InputPinV0 for Pin<mode::OpenDrain, PIN> {
     }
 }
 
-impl<PIN: PinOps> ErrorType for Pin<mode::OpenDrain, PIN> { type Error = core::convert::Infallible; }
+impl<PIN: PinOps> ErrorType for Pin<mode::OpenDrain, PIN> {
+    type Error = core::convert::Infallible;
+}
 
 impl<PIN: PinOps> InputPin for Pin<mode::OpenDrain, PIN> {
     fn is_high(&mut self) -> Result<bool, Self::Error> {
@@ -464,7 +469,9 @@ impl<PIN: PinOps, IMODE: mode::InputMode> InputPinV0 for Pin<mode::Input<IMODE>,
     }
 }
 
-impl<PIN: PinOps, IMODE: mode::InputMode> ErrorType for Pin<mode::Input<IMODE>, PIN> { type Error = core::convert::Infallible; }
+impl<PIN: PinOps, IMODE: mode::InputMode> ErrorType for Pin<mode::Input<IMODE>, PIN> {
+    type Error = core::convert::Infallible;
+}
 
 impl<PIN: PinOps, IMODE: mode::InputMode> InputPin for Pin<mode::Input<IMODE>, PIN> {
     fn is_high(&mut self) -> Result<bool, Self::Error> {
@@ -529,6 +536,27 @@ impl<PIN: PinOps> Pin<mode::Analog, PIN> {
     {
         crate::adc::Channel::new(self)
     }
+
+    /// Convert this pin to a floating digital input pin.
+    ///
+    /// The pin is re-enabled in the digital input buffer and is no longer usable as an analog
+    /// input. You can get to other digital modes by calling one of the usual `into_...` methods
+    /// on the return value of this function.
+    pub fn into_digital<H, ADC, CLOCK>(
+        self,
+        adc: &mut crate::adc::Adc<H, ADC, CLOCK>,
+    ) -> Pin<mode::Input<mode::Floating>, PIN>
+    where
+        Pin<mode::Analog, PIN>: crate::adc::AdcChannel<H, ADC>,
+        ADC: crate::adc::AdcOps<H>,
+        CLOCK: crate::clock::Clock,
+    {
+        adc.disable_pin(&self);
+        Pin {
+            pin: self.pin,
+            _mode: PhantomData,
+        }
+    }
 }
 
 #[macro_export]
@@ -584,10 +612,10 @@ macro_rules! impl_port_traditional {
         }
 
         impl Dynamic {
-            fn new(port: DynamicPort, pin_num: u8) -> Self {
+            fn new(port: DynamicPort, num: u8) -> Self {
                 Self {
                     port,
-                    mask: 1 << pin_num,
+                    mask: 1u8 << num,
                 }
             }
         }
@@ -622,7 +650,7 @@ macro_rules! impl_port_traditional {
                 #[inline]
                 unsafe fn out_toggle(&mut self) {
                     match self.port {
-                        $(DynamicPort::[<PORT $name>] => (*<$port>::ptr()).[<port $name:lower>].write(|w| {
+                        $(DynamicPort::[<PORT $name>] => (*<$port>::ptr()).[<pin $name:lower>].write(|w| {
                             w.bits(self.mask)
                         }),)+
                     }
@@ -631,16 +659,18 @@ macro_rules! impl_port_traditional {
                 #[inline]
                 unsafe fn out_get(&self) -> bool {
                     match self.port {
-                        $(DynamicPort::[<PORT $name>] => (*<$port>::ptr()).[<port $name:lower>].read().bits()
-                            & self.mask != 0,)+
+                        $(DynamicPort::[<PORT $name>] => {
+                            (*<$port>::ptr()).[<port $name:lower>].read().bits() & self.mask != 0
+                        })+
                     }
                 }
 
                 #[inline]
                 unsafe fn in_get(&self) -> bool {
                     match self.port {
-                        $(DynamicPort::[<PORT $name>] => (*<$port>::ptr()).[<pin $name:lower>].read().bits()
-                            & self.mask != 0,)+
+                        $(DynamicPort::[<PORT $name>] => {
+                            (*<$port>::ptr()).[<pin $name:lower>].read().bits() & self.mask != 0
+                        })+
                     }
                 }
 
@@ -685,44 +715,46 @@ macro_rules! impl_port_traditional {
 
                     #[inline]
                     unsafe fn out_set(&mut self) {
-                        (*<$port>::ptr()).[<port $name:lower>].modify(|r, w| {
-                            w.bits(r.bits() | (1 << $pin))
+                        (*<$port>::ptr()).[<port $name:lower>].modify(|_, w| {
+                            w.[<p $name:lower $pin>]().set_bit()
                         })
                     }
 
                     #[inline]
                     unsafe fn out_clear(&mut self) {
-                        (*<$port>::ptr()).[<port $name:lower>].modify(|r, w| {
-                            w.bits(r.bits() & !(1 << $pin))
+                        (*<$port>::ptr()).[<port $name:lower>].modify(|_, w| {
+                            w.[<p $name:lower $pin>]().clear_bit()
                         })
                     }
 
                     #[inline]
                     unsafe fn out_toggle(&mut self) {
-                        (*<$port>::ptr()).[<pin $name:lower>].write(|w| w.bits(1 << $pin))
+                        (*<$port>::ptr()).[<pin $name:lower>].write(|w| {
+                            w.[<p $name:lower $pin>]().set_bit()
+                        })
                     }
 
                     #[inline]
                     unsafe fn out_get(&self) -> bool {
-                        (*<$port>::ptr()).[<port $name:lower>].read().bits() & (1 << $pin) != 0
+                        (*<$port>::ptr()).[<port $name:lower>].read().[<p $name:lower $pin>]().bit()
                     }
 
                     #[inline]
                     unsafe fn in_get(&self) -> bool {
-                        (*<$port>::ptr()).[<pin $name:lower>].read().bits() & (1 << $pin) != 0
+                        (*<$port>::ptr()).[<pin $name:lower>].read().[<p $name:lower $pin>]().bit()
                     }
 
                     #[inline]
                     unsafe fn make_output(&mut self) {
-                        (*<$port>::ptr()).[<ddr $name:lower>].modify(|r, w| {
-                            w.bits(r.bits() | (1 << $pin))
+                        (*<$port>::ptr()).[<ddr $name:lower>].modify(|_, w| {
+                            w.[<p $name:lower $pin>]().set_bit()
                         })
                     }
 
                     #[inline]
                     unsafe fn make_input(&mut self, pull_up: bool) {
-                        (*<$port>::ptr()).[<ddr $name:lower>].modify(|r, w| {
-                            w.bits(r.bits() & !(1 << $pin))
+                        (*<$port>::ptr()).[<ddr $name:lower>].modify(|_, w| {
+                            w.[<p $name:lower $pin>]().clear_bit()
                         });
                         if pull_up {
                             self.out_set()