+++ mcu/atmega-hal/src/port.rs
@@ -9,11 +9,21 @@ avr_hal_generic::impl_port_traditional! {
     }
 }
 
+#[cfg(any(feature = "atmega164pa"))]
+avr_hal_generic::impl_port_traditional! {
+    enum Ports {
+        A: crate::pac::PORTA = [0, 1, 2, 3, 4, 5, 6 ,7],
+        B: crate::pac::PORTB = [0, 1, 2, 3, 4, 5, 6 ,7],
+        C: crate::pac::PORTC = [0, 1, 2, 3, 4, 5, 6 ,7],
+        D: crate::pac::PORTD = [0, 1, 2, 3, 4, 5, 6 ,7],
+    }
+}
+
 #[cfg(feature = "atmega328pb")]
 avr_hal_generic::impl_port_traditional! {
     enum Ports {
         B: crate::pac::PORTB = [0, 1, 2, 3, 4, 5, 6, 7],
-        C: crate::pac::PORTC = [0, 1, 2, 3, 4, 5, 6, 7],
+        C: crate::pac::PORTC = [0, 1, 2, 3, 4, 5, 6],
         D: crate::pac::PORTD = [0, 1, 2, 3, 4, 5, 6, 7],
         E: crate::pac::PORTE = [0, 1, 2, 3],
     }
@@ -26,7 +36,7 @@ avr_hal_generic::impl_port_traditional! {
         C: crate::pac::PORTC = [6, 7],
         D: crate::pac::PORTD = [0, 1, 2, 3, 4, 5, 6, 7],
         E: crate::pac::PORTE = [2, 6],
-        F: crate::pac::PORTF = [0, 1, 2, 3, 4, 5, 6, 7],
+        F: crate::pac::PORTF = [0, 1, 4, 5, 6, 7],
     }
 }
 
@@ -39,7 +49,7 @@ avr_hal_generic::impl_port_traditional! {
         D: crate::pac::PORTD = [0, 1, 2, 3, 4, 5, 6, 7],
         E: crate::pac::PORTE = [0, 1, 2, 3, 4, 5, 6, 7],
         F: crate::pac::PORTF = [0, 1, 2, 3, 4, 5, 6, 7],
-        G: crate::pac::PORTG = [0, 1, 2, 3, 4, 5],
+        G: crate::pac::PORTG = [0, 1, 2, 3, 4],
     }
 }