+++ mcu/atmega-hal/src/spi.rs
@@ -2,7 +2,12 @@
 use crate::port;
 pub use avr_hal_generic::spi::*;
 
-#[cfg(any(feature = "atmega128a", feature = "atmega1280", feature = "atmega2560", feature = "atmega32u4"))]
+#[cfg(any(
+    feature = "atmega128a",
+    feature = "atmega1280",
+    feature = "atmega2560",
+    feature = "atmega32u4"
+))]
 pub type Spi = avr_hal_generic::spi::Spi<
     crate::Atmega,
     crate::pac::SPI,
@@ -11,7 +16,12 @@ pub type Spi = avr_hal_generic::spi::Spi<
     port::PB3,
     port::PB0,
 >;
-#[cfg(any(feature = "atmega128a", feature = "atmega1280", feature = "atmega2560", feature = "atmega32u4"))]
+#[cfg(any(
+    feature = "atmega128a",
+    feature = "atmega1280",
+    feature = "atmega2560",
+    feature = "atmega32u4"
+))]
 avr_hal_generic::impl_spi! {
     hal: crate::Atmega,
     peripheral: crate::pac::SPI,
@@ -21,7 +31,12 @@ avr_hal_generic::impl_spi! {
     cs: port::PB0,
 }
 
-#[cfg(any(feature = "atmega168", feature = "atmega328p", feature = "atmega48p",))]
+#[cfg(any(
+    feature = "atmega168",
+    feature = "atmega328p",
+    feature = "atmega48p",
+    feature = "atmega8"
+))]
 pub type Spi = avr_hal_generic::spi::Spi<
     crate::Atmega,
     crate::pac::SPI,
@@ -30,7 +45,12 @@ pub type Spi = avr_hal_generic::spi::Spi<
     port::PB4,
     port::PB2,
 >;
-#[cfg(any(feature = "atmega168", feature = "atmega328p", feature = "atmega48p",))]
+#[cfg(any(
+    feature = "atmega168",
+    feature = "atmega328p",
+    feature = "atmega48p",
+    feature = "atmega8"
+))]
 avr_hal_generic::impl_spi! {
     hal: crate::Atmega,
     peripheral: crate::pac::SPI,
@@ -94,14 +114,4 @@ avr_hal_generic::impl_spi! {
     mosi: port::PB5,
     miso: port::PB6,
     cs: port::PB4,
-}
-
-#[cfg(any(feature = "atmega8"))]
-avr_hal_generic::impl_spi! {
-    hal: crate::Atmega,
-    peripheral: crate::pac::SPI,
-    sclk: port::PB5,
-    mosi: port::PB3,
-    miso: port::PB4,
-    cs: port::PB2,
 }